send more info back with lists, show first initial name in lists, debug round trip

+++ src/main/java/mitll/langtest/server/database/result/AnswerDAO.java
@@ -243,7 +243,7 @@ public class AnswerDAO extends BaseAnswerDAO implements IAnswerDAO {
   }
 
   @Override
-  public void addRoundTrip(int resultID, int roundTrip) {
+  public void addRoundTrip(int resultID, long roundTrip) {
     Connection connection = getConnection();
     try {
       String sql = "UPDATE " +
